refactor(briefing): modularize BriefingDashboard into components/hooks/lib
- Extract TOC + markdown logic to lib/briefing-toc.ts (pure, unit-tested)
- Extract data fetching to lib/hooks/useBriefingData.ts
- Split UI into components/briefing/{BriefingPanel,BriefingRow,BriefingToc,icons}
- Reduce inline SVG duplication via shared icon components
- BriefingDashboard.tsx is now a thin 102-line orchestrator (was 507)
- Add lib/briefing-toc unit tests; all 143 web tests pass
